+++ packages/traceloop-sdk/tests/conftest.py
@@ -1,4 +1,6 @@
 """Unit tests configuration module."""
+
+import os
 import pytest
 from traceloop.sdk import Traceloop
 from opentelemetry.sdk.trace.export.in_memory_span_exporter import InMemorySpanExporter
@@ -21,3 +23,13 @@ def exporter():
 @pytest.fixture(autouse=True)
 def clear_exporter(exporter):
     exporter.clear()
+
+
+@pytest.fixture(autouse=True)
+def environment():
+    os.environ["OPENAI_API_KEY"] = "test_api_key"
+
+
+@pytest.fixture(scope="module")
+def vcr_config():
+    return {"filter_headers": ["authorization"]}